template_source_code/requirements.txt
json5==0.9.6
numpy
flask==2.0.3
Flask-Cors
werkzeug==2.0.3
geopandas==0.11.0
psycopg2-binary
SQLAlchemy==1.4.20
